health-monitor-main/requirements.txt
Flask>=2.3.2
PyYAML>=6.0
docker>=6.1.3
kubernetes>=26.1.0
watchdog>=3.0.0
psutil==5.9.5
colorama==0.4.6
rich==13.6.0